Fix face swap V3 pipeline: distilled single-stage, bfloat16 preprocess crash.
Align FaceSwapPipeline with Comfy v3 workflow (full-res distilled 8-step,
no IC-LoRA guide normalize). Fix mlx bfloat16→numpy failure in guide
preprocess and return float32 from PyAV video loader.

+++ ltx_media.py
@@ -187,7 +187,7 @@ def load_video_frames_normalized(
     stacked = np.stack(frames_list, axis=0)
     tensor = mx.array(stacked).transpose(0, 3, 1, 2)
     tensor = tensor.transpose(1, 0, 2, 3)[None, ...]
-    return tensor.astype(mx.bfloat16)
+    return tensor.astype(mx.float32)
 
 
 def probe_audio_duration(path: Path | str) -> float | None:
